refactor conftest
