+++ tests/conftest.py
@@ -1,0 +1,16 @@
+import pytest
+
+
+@pytest.fixture(scope="session")
+def acc(accounts):
+    yield accounts[0]
+
+
+@pytest.fixture(scope="session")
+def alice(accounts):
+    yield accounts[1]
+
+
+@pytest.fixture(scope="session")
+def bob(accounts):
+    yield accounts[2]
