+++ tests/contracts/conftest.py
@@ -4,21 +4,6 @@ from ape import Contract, project
 
 
 @pytest.fixture(scope="module")
-def acc(accounts):
-    yield accounts[0]
-
-
-@pytest.fixture(scope="module")
-def alice(accounts):
-    yield accounts[1]
-
-
-@pytest.fixture(scope="module")
-def bob(accounts):
-    yield accounts[2]
-
-
-@pytest.fixture(scope="module")
 def backtest(acc):
     yield project.MockBacktest.deploy(sender=acc)
