--- tests/scripts/examples/uniswap/v3/conftest.py
@@ -1,1 +1,0 @@
-
